full stack project/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.1
pymysql==1.1.0
python-pptx==0.6.21
